Rename vector_scope_thread -> vector_scope
