+++ src-tauri/src/model.rs
@@ -1,5 +1,5 @@
 pub mod graph_plotter;
 pub mod mouse_info;
 pub mod screenshot_capture;
-pub mod vector_scope_thread;
+pub mod vector_scope;
 pub mod worker_thread;
