+++ src-tauri/src/model/vector_scope.rs
